requirements.txt
langchain-core
langchain-ollama
langchain-agents
langgraph

requests
python-dotenv